+++ src/remove.js
@@ -1,7 +1,7 @@
 import defaultSchema from './utils/default-schema';
 import idFromResource from './utils/id-from-resource';
 import merge from './utils/merge';
-import { exists, isArray, isObject, isNull } from './utils/identification';
+import { isArray, isObject, isNull } from './utils/identification';
 import { warning } from './utils/warning';
 import createChanges from './utils/create-changes';
 
@@ -14,8 +14,6 @@ import createChanges from './utils/create-changes';
 // });
 
 export default function remove({ path, schemas, state, changes }) {
-  const newState = merge(state);
-
   changes = createChanges(path, changes);
 
   if (!isObject(changes) && process.env.NODE_ENV !== 'production') {
@@ -26,120 +24,104 @@ export default function remove({ path, schemas, state, changes }) {
     );
   }
 
-  for (let resourceType in changes) {
-    const resourceChange = changes[resourceType];
-    const currentResourceSection = state[resourceType];
+  const resourcesChanges = changes.resources;
+  const listsChanges = changes.lists;
+  const newResources = merge(state.resources);
+
+  const resourcesDeletedByType = {};
+
+  for (let resourceType in resourcesChanges) {
+    const currentResourceSection = newResources[resourceType];
 
     if (!currentResourceSection) {
       continue;
     }
 
-    if (process.env.NODE_ENV !== 'production') {
-      if (exists(resourceChange) && !isObject(resourceChange)) {
-        warning(
-          `You called remove with an invalid update for the` +
-            ` resource of type "${resourceType}". Updates must be an Object.`,
-          'DELETE_RESOURCES_INVALID_TYPE',
-          'error'
-        );
-      } else {
-        if (exists(resourceChange.resources)) {
-          if (
-            !isObject(resourceChange.resources) &&
-            !isArray(resourceChange.resources)
-          ) {
-            warning(
-              `You called remove with an invalid "resources" value for the` +
-                ` resource of type "${resourceType}". The "resource" value of an update` +
-                ` must be an Object or an array.`,
-              'DELETE_RESOURCES_INVALID_RESOURCES',
-              'error'
-            );
-          }
-        }
+    resourcesDeletedByType[resourceType] = {};
 
-        if (exists(resourceChange.lists)) {
-          if (
-            !isArray(resourceChange.lists) &&
-            !isObject(resourceChange.lists)
-          ) {
-            warning(
-              `You called remove with an invalid "lists" value for the` +
-                ` resource of type "${resourceType}". The "lists" value when deleting` +
-                ` must be an array or an object.`,
-              'DELETE_RESOURCES_INVALID_LISTS',
-              'error'
-            );
-          }
-        }
-      }
-    }
-
-    const listIsArray = isArray(resourceChange.lists);
-    const listIsObject = isObject(resourceChange.lists);
-
-    let listsToUse;
-    if (listIsArray || listIsObject) {
-      listsToUse = resourceChange.lists;
-    } else {
-      listsToUse = [];
-    }
-
-    let listsToDelete = [];
-    if (listIsArray) {
-      listsToDelete = listsToUse;
-    } else if (listIsObject) {
-      for (let listName in listsToUse) {
-        if (isNull(listsToUse[listName])) {
-          listsToDelete.push(listName);
-        }
-      }
-      listsToDelete;
-    }
-
-    const naiveResources = resourceChange && resourceChange.resources;
+    const resourceChange = resourcesChanges[resourceType];
     const schema = schemas[resourceType] || defaultSchema;
 
     let idList = [];
-    if (isArray(naiveResources)) {
-      idList = naiveResources.map(resource =>
-        idFromResource({ schema, resource })
-      );
-    }
 
-    const hasIds = idList && idList.length;
-    const newResources = merge(currentResourceSection.resources);
+    if (isArray(resourceChange)) {
+      idList = resourceChange.map(resource => {
+        const id = idFromResource({ schema, resource });
 
-    if (hasIds) {
-      idList.map(id => {
-        delete newResources[id];
+        resourcesDeletedByType[resourceType][id] = true;
+
+        return id;
       });
     }
 
-    let newLists = {};
-    const lists = currentResourceSection.lists;
+    const hasIds = idList && idList.length;
+    const newResourceSection = merge(currentResourceSection);
 
-    // We iterate every existing list
-    for (let resourceList in lists) {
-      // We only consider the lists that aren't slated to be deleted
-      if (!listsToDelete.includes(resourceList)) {
-        const existingList = lists[resourceList];
-        const thisListFromChange = listIsObject ? listsToUse[resourceList] : [];
-        const thisListIsArray = isArray(thisListFromChange);
-        newLists[resourceList] = existingList.filter(r => {
-          const removedFromStore = idList.includes(r);
-          const removeFromThisList =
-            thisListIsArray && thisListFromChange.includes(r);
-          return !removedFromStore && !removeFromThisList;
-        });
-      }
+    if (hasIds) {
+      idList.map(id => {
+        delete newResourceSection[id];
+      });
     }
 
-    newState[resourceType] = merge(currentResourceSection, {
-      resources: newResources,
-      lists: newLists,
-    });
+    newResources[resourceType] = newResourceSection;
   }
 
-  return newState;
+  const listIsArray = isArray(listsChanges);
+  const listIsObject = isObject(listsChanges);
+
+  let listsToUse;
+  if (listIsArray || listIsObject) {
+    listsToUse = listsChanges;
+  } else {
+    listsToUse = [];
+  }
+
+  let listsToDelete = [];
+  if (listIsArray) {
+    listsToDelete = listsToUse;
+  } else if (listIsObject) {
+    for (let listName in listsToUse) {
+      if (isNull(listsToUse[listName])) {
+        listsToDelete.push(listName);
+      }
+    }
+    listsToDelete;
+  }
+
+  const newLists = {};
+
+  // We iterate every existing list
+  for (let resourceList in state.lists) {
+    // We only consider the lists that aren't slated to be deleted
+    if (!listsToDelete.includes(resourceList)) {
+      const existingList = state.lists[resourceList];
+      const thisListFromChange = listIsObject ? listsToUse[resourceList] : [];
+      const thisListIsArray = isArray(thisListFromChange);
+
+      newLists[resourceList] = existingList.filter(resourcePointer => {
+        const schema = schemas[resourcePointer.resourceType] || defaultSchema;
+        const deletedType =
+          resourcesDeletedByType[resourcePointer.resourceType] || {};
+        const removedFromStore =
+          deletedType[resourcePointer[schema.idProperty]];
+
+        const removeFromThisList =
+          thisListIsArray &&
+          thisListFromChange.find(resource => {
+            return (
+              resource[schema.idProperty] ===
+                resourcePointer[schema.idProperty] &&
+              resource.resourceType === resourcePointer.resourceType
+            );
+          });
+
+        return !removedFromStore && !removeFromThisList;
+      });
+    }
+  }
+
+  return merge(state, {
+    resources: newResources,
+    lists: newLists,
+  });
 }
